Contao 4 changes
