+++ composer.json
@@ -20,22 +20,15 @@
 		"source":"https://github.com/contao-community-alliance/events-cron"
 	},
 	"require":{
-		"php":">=5.3",
-		"contao/core": ">=2.11,<4.0-dev",
-		"contao-community-alliance/composer-plugin": "~2.0",
-		"contao-community-alliance/event-dispatcher": "~1.0"
+		"php":"^5.6|^7.0",
+		"contao/core-bundle": "~4.4"
 	},
 	"autoload":{
-		"psr-0":{
-			"ContaoCommunityAlliance\\Contao\\Events\\Cron": "src/"
+		"psr-4":{
+			"ContaoCommunityAlliance\\Contao\\Events\\Cron\\": "src/"
 		}
 	},
 	"extra": {
-		"contao": {
-			"sources": {
-				"contao": "system/modules/events-cron"
-			}
-		},
 		"branch-alias": {
 			"dev-master": "1.0.x-dev",
 			"dev-develop": "1.1.x-dev"
